feat: add CFUQuestion and CFUQuestionStudentAnswer components for enhanced question display
